+++ CHANGELOG.md
@@ -22,7 +22,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Transportalizers now stack to one instead of 64
 - Prismarine armor remodel and retexture
 - Iron Lass Armor has now been remodeled, and retextured; with additional new animations.
-
+- Computer Themes can now be added and rendered via resource pack
+- Default computer Theme for crockertop is now Crocker
+- Computer Theme is now chosen from a list of available Themes in separate GUI
 
 ### Fixed